+++ internal/service/database/database.go
@@ -412,14 +412,26 @@ func resourceUpCloudManagedDatabaseSetCommonState(d *schema.ResourceData, detail
 	return d.Set("primary_database", details.ServiceURIParams.DatabaseName)
 }
 
-func getDatabaseDeleted(ctx context.Context, svc *service.Service, id ...string) (map[string]interface{}, error) {
-	db, err := svc.GetManagedDatabase(ctx, &request.GetManagedDatabaseRequest{UUID: id[0]})
+func getDatabaseDetails(ctx context.Context, svc *service.Service, id ...string) (*utils.ResourceDetails, error) {
+	db, err := svc.GetManagedDatabase(ctx, &request.GetManagedDatabaseRequest{
+		UUID: id[0],
+	})
 
-	return map[string]interface{}{"resource": "database", "name": db.Name, "state": db.State}, err
+	details := &utils.ResourceDetails{
+		ResourceType: "database",
+	}
+
+	if db != nil {
+		details.Name = db.Name
+		details.State = string(db.State)
+		details.Running = db.State == upcloud.ManagedDatabaseStateRunning
+	}
+
+	return details, err
 }
 
 func waitForDatabaseToBeDeleted(ctx context.Context, svc *service.Service, id string) error {
-	return utils.WaitForResourceToBeDeleted(ctx, svc, getDatabaseDeleted, id)
+	return utils.WaitForResourceToBeDeleted(ctx, svc, getDatabaseDetails, id)
 }
 
 func waitServiceNameToPropagate(ctx context.Context, name string) (err error) {